Created Scraping Program for Toronto Parks Data
Created an automated scraper that Uses a json file which contains Toronto parks: IDs, Names, Latitude, and Longitude. Each individual page is visited by swapping the park ID in the URL. The page is then checked using Nokogiri methods so see if the park has facilities for relevant sports to this app. Parks with relevant facilities are added to an array. The array is written to JSON format when the final park has been scraped.