Add new users with password_digest
New User post requests from react front end now add a user to the db when validation is passed. There was an error with our bcrypt install so a workaround function was added to give the user a hash password. The functionality of the work around was verified by creating a user authenticate function and testing in byebug.