Preferences Controller and Time Prefs
Created a new controller for user preferences. Created the Show method which pulls all preferences for a user from the database and organizes the data so it is easy to use on the front end.

Edited the timeprefs table to include a boolean column named active. When a new user is created, they will have a row inserted in this table for every day of the week with active set to false. When a user adds their first timepref, active will change to true and times will be added.